move src files to docs
