+++ docs/assets/algo-filipino-logo.png
