+++ docs/assets/delete-this.png
