+++ docs/assets/event-placeholder-1.png
