+++ docs/assets/event-placeholder-2.png
