+++ docs/assets/event-training.png
