+++ docs/assets/event-workshop.png
