+++ docs/assets/recent-placeholder-1.png
