+++ docs/assets/recent-placeholder-2.png
