+++ docs/components/Footer.jsx
