--- src/App.css
@@ -1,8 +1,0 @@
-@tailwind base;
-@tailwind components;
-@tailwind utilities;
-
-* {
-  margin: 0px;
-  padding: 0px;
-}
