--- src/App.js
@@ -1,33 +1,0 @@
-import React from "react";
-import Navbar from "./components/Navbar";
-import Header from "./components/Header";
-import { BrowserRouter as Router, Switch, Route } from "react-router-dom";
-import "./App.css";
-import "tailwindcss/tailwind.css"
-import Introduction from "./components/Introduction";
-import Testimonials from "./components/Testimonials";
-import JoinUs from "./components/JoinUs"; 
-import Footer from "./components/Footer";
-
-function App() {
-  return (
-    <>
-      <Router>
-        <Navbar />
-        <div className="px-6 lg:px-36">
-          <Header />
-          <Introduction />
-          <Testimonials />
-          <JoinUs />
-        </div>
-        <Footer />
-
-        <Switch>
-          <Route path="/" exact />
-        </Switch>
-      </Router>
-    </>
-  );
-}
-
-export default App;
